CRUD service of vendor version Beta

+++ app/src/main/java/com/example/androiddevelopmentgroup7/dataModels/Service.kt
@@ -8,6 +8,6 @@ data class Service(
     var serviceContact:String,
     var serviceImage:String,
 ){
-    var serviceRating:Double = 0.0
+    var serviceRating:Float = 0.toFloat()
 }
